restructure the files so that each driver can be built individually
