+++ mocap_base/package.xml
@@ -1,21 +1,17 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>motion_capture_system</name>
+  <name>mocap_base</name>
   <version>0.0.1</version>
-  <description>A unifying folder for different motion capture systems</description>
+  <description>Base class for different motion capture systems</description>
 
   <maintainer email="sunke.polyu@gmail.com">Ke Sun</maintainer>
   <license>Apache-2.0</license>
   <author>Ke Sun</author>
 
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>message_generation</build_depend>
-  <exec_depend>message_runtime</exec_depend>
   <depend>geometry_msgs</depend>
   <depend>roscpp</depend>
-  <depend>tf</depend>
   <depend>eigen_conversions</depend>
-  <depend>tf_conversions</depend>
 
   <export>
   </export>
