+++ mocap_qualisys/include/mocap_qualisys/StdString.h
