+++ mocap_qualisys/src/QualisysDriver.cpp
@@ -20,7 +20,7 @@
 #include <Eigen/Dense>
 #include <Eigen/Geometry>
 #include <tf_conversions/tf_eigen.h>
-#include <motion_capture_system/qualisys/QualisysDriver.h>
+#include <mocap_qualisys/QualisysDriver.h>
 
 using namespace std;
 using namespace Eigen;
