+++ mocap_vicon/CMakeLists.txt
@@ -1,22 +1,21 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(motion_capture_system)
+project(mocap_vicon)
 
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_CURRENT_LIST_DIR}/cmake")
 
 find_package(Boost REQUIRED)
 find_package(Eigen3 REQUIRED)
 find_package(catkin REQUIRED COMPONENTS
-  message_generation
   roscpp
   geometry_msgs
   tf
-  eigen_conversions
   tf_conversions
+  mocap_base
 )
 
 if(NOT APPLE)
   if(CMAKE_SIZEOF_VOID_P EQUAL 8)
-    set(SDK_DIR "${PROJECT_SOURCE_DIR}/include/${PROJECT_NAME}/vicon/vicon_sdk/Linux64")
+    set(SDK_DIR "${PROJECT_SOURCE_DIR}/include/${PROJECT_NAME}/vicon_sdk/Linux64")
     set(VICON_BOOST_LIB_SUFFIX ".so.1.51.0")
     list(APPEND CMAKE_FIND_LIBRARY_SUFFIXES ${VICON_BOOST_LIB_SUFFIX})
     find_library(VICON_BOOST_THREAD_LIB boost_thread-mt ${SDK_DIR}
@@ -30,7 +29,7 @@ if(NOT APPLE)
     set(VICON_BOOST_LIBS ${VICON_BOOST_THREAD_LIB} ${VICON_BOOST_SYSTEM_LIB}
       ${VICON_BOOST_CHRONO_LIB})
   else(CMAKE_SIZEOF_VOID_P EQUAL 8)
-    set(SDK_DIR "${PROJECT_SOURCE_DIR}/include/${PROJECT_NAME}/vicon/vicon_sdk/Linux32")
+    set(SDK_DIR "${PROJECT_SOURCE_DIR}/include/${PROJECT_NAME}/icon_sdk/Linux32")
     set(VICON_BOOST_LIB_SUFFIX ".so.1.51.0")
     list(APPEND CMAKE_FIND_LIBRARY_SUFFIXES ${VICON_BOOST_LIB_SUFFIX})
     find_library(VICON_BOOST_WAVE_LIB boost_wave-mt ${SDK_DIR} NO_DEFAULT_PATH)
@@ -55,20 +54,18 @@ if(NOT APPLE)
   #message(STATUS "SDK_DIR: ${SDK_DIR}")
   #message(STATUS "VICON_BOOST_LIBS: ${VICON_BOOST_LIBS}")
 else(NOT APPLE)
-  set(SDK_DIR "${PROJECT_SOURCE_DIR}/include/${PROJECT_NAME}/vicon/vicon_sdk/MacOSX")
+  set(SDK_DIR "${PROJECT_SOURCE_DIR}/include/${PROJECT_NAME}/vicon_sdk/MacOSX")
 endif()
 
 find_library(VICON_SDK_LIB ViconDataStreamSDK_CPP ${SDK_DIR} NO_DEFAULT_PATH)
 find_library(VICON_DEBUG_SERVICES_LIB DebugServices ${SDK_DIR} NO_DEFAULT_PATH)
 
-#add_message_files(DIRECTORY msg)
-#generate_messages(DEPENDENCIES geometry_msgs)
 
 catkin_package(
   #INCLUDE_DIRS include
   #LIBRARIES ${PROJECT_NAME}
   CATKIN_DEPENDS
-    message_runtime geometry_msgs tf
+    geometry_msgs tf mocap_base
   DEPENDS
   Boost Eigen3
 )
@@ -82,27 +79,14 @@ include_directories(
 
 add_definitions("-std=c++0x -Wall")
 
-add_library(mocap_kalman_filter
-  src/KalmanFilter.cpp
-)
-target_link_libraries(mocap_kalman_filter
-  ${catkin_LIBRARIES}
-)
-add_dependencies(mocap_kalman_filter
-  ${${PROJECT_NAME}_EXPORTED_TARGETS}
-  ${catkin_EXPORTED_TARGETS}
-)
 
-# VICON
 add_library(mocap_vicon_driver
-  src/MoCapDriverBase.cpp
-  src/vicon/ViconDriver.cpp
+  src/ViconDriver.cpp
 )
 target_link_libraries(mocap_vicon_driver
   ${VICON_SDK_LIB}
   ${VICON_DEBUG_SERVICES_LIB}
   ${VICON_BOOST_LIBS}
-  mocap_kalman_filter
   ${catkin_LIBRARIES}
 )
 add_dependencies(mocap_vicon_driver
@@ -111,37 +95,9 @@ add_dependencies(mocap_vicon_driver
 )
 
 add_executable(mocap_vicon_node
-  src/vicon/vicon.cpp
+  src/vicon.cpp
 )
 target_link_libraries(mocap_vicon_node
   mocap_vicon_driver
   ${catkin_LIBRARIES}
 )
-
-# QUALISYS
-add_library(mocap_qualisys_driver
-  src/MoCapDriverBase.cpp
-  src/qualisys/QualisysDriver.cpp
-  include/motion_capture_system/qualisys/RTProtocol.cpp
-  include/motion_capture_system/qualisys/NBC_Markup.cpp
-  include/motion_capture_system/qualisys/RTPacket.cpp
-  include/motion_capture_system/qualisys/Network.cpp
-)
-target_link_libraries(mocap_qualisys_driver
-  mocap_kalman_filter
-  ${catkin_LIBRARIES}
-)
-add_dependencies(mocap_qualisys_driver
-  ${${PROJECT_NAME}_EXPORTED_TARGETS}
-
-  ${catkin_EXPORTED_TARGETS}
-)
-
-add_executable(mocap_qualisys_node
-  src/qualisys/qualisys.cpp
-)
-target_link_libraries(mocap_qualisys_node
-  mocap_qualisys_driver
-  ${catkin_LIBRARIES}
-)
-
