+++ mocap_vicon/include/mocap_vicon/vicon_sdk/Linux32/Client.h
