+++ mocap_vicon/launch/vicon.launch
@@ -1,7 +1,7 @@
 <launch>
-  <node pkg="motion_capture_system"
+  <node pkg="mocap_vicon"
     type="mocap_vicon_node"
-    name="vicon_new"
+    name="vicon"
     output="screen">
     <param name="vicon_server" value="alkaline2"/>
     <param name="frame_rate" value="100"/>
