+++ mocap_vicon/package.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>mocap_vicon</name>
+  <version>0.0.1</version>
+  <description>Driver for Vicon</description>
+
+  <maintainer email="sunke.polyu@gmail.com">Ke Sun</maintainer>
+  <license>Apache-2.0</license>
+  <author>Ke Sun</author>
+
+  <buildtool_depend>catkin</buildtool_depend>
+
+  <depend>geometry_msgs</depend>
+  <depend>roscpp</depend>
+  <depend>tf</depend>
+  <depend>tf_conversions</depend>
+  <depend>mocap_base</depend>
+
+  <export>
+  </export>
+</package>
